+++ app/src/main/java/com/example/weatherapp/ui/notifications/NotificationsFragment.java
@@ -2,10 +2,14 @@ package com.example.weatherapp.ui.notifications;
 
 import android.content.Intent;
 import android.net.Uri;
+import android.os.Build;
 import android.os.Bundle;
+import android.os.Handler;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.webkit.WebView;
 import android.widget.SeekBar;
 import android.widget.TextView;
 
@@ -19,8 +23,21 @@ import androidx.lifecycle.ViewModelProviders;
 import com.example.weatherapp.CustomTextView;
 import com.example.weatherapp.R;
 
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.util.stream.Collectors;
+
+import javax.net.ssl.HttpsURLConnection;
+
 public class NotificationsFragment extends Fragment {
     CustomTextView customTextView;
+    private String url = "https://www.google.ru";
+    WebView webView;
+    SeekBar seekBar;
+
 
     private NotificationsViewModel notificationsViewModel;
 
@@ -37,11 +54,31 @@ public class NotificationsFragment extends Fragment {
             @Override
             public void onChanged(@Nullable String s) {
                 textView.setText(s);
-
-
             }
         });
 
+        webView = root.findViewById(R.id.webView);
+        seekBar = root.findViewById(R.id.seekBar);
+
+        seekBar.setOnSeekBarChangeListener(new SeekBar.OnSeekBarChangeListener() {
+            @Override
+            public void onProgressChanged(SeekBar seekBar, int progress, boolean fromUser) {
+                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+                    webView.zoomBy(progress);
+                }
+            }
+
+            @Override
+            public void onStartTrackingTouch(SeekBar seekBar) {
+
+            }
+
+            @Override
+            public void onStopTrackingTouch(SeekBar seekBar) {
+
+            }
+        });
+//        webView.loadUrl(url);
 
 
         return root;
@@ -62,7 +99,55 @@ public class NotificationsFragment extends Fragment {
     @Override
     public void onActivityCreated(@Nullable Bundle savedInstanceState) {
         super.onActivityCreated(savedInstanceState);
+        loadWedView(webView);
 
 
+
+    }
+
+    private void loadWedViewRetrofit(final WebView webView) {
+
+    }
+
+
+        private void loadWedView(final WebView webView) {
+        final Handler handler = new Handler();
+        Log.i("Thread", Thread.currentThread().getName());
+        new Thread(new Runnable() {
+            @Override
+            public void run() {
+                try {
+                    Log.i("Thread", Thread.currentThread().getName());
+
+                    URL uri = new URL(url);
+                    HttpsURLConnection connection = (HttpsURLConnection) uri.openConnection();
+                    connection.setRequestMethod("GET");
+                    connection.setConnectTimeout(10000);
+                    connection.connect();
+
+                        handler.post(new Runnable() {
+                            @Override
+                            public void run() {
+                                webView.loadUrl(url);
+//                                webView.loadData(result, "text/html; charset=utf-8", "utf-8");
+                            }
+                        });
+
+                    BufferedReader in = new BufferedReader(new InputStreamReader(connection.getInputStream()));
+                    if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.N) {
+                        final String result = in.lines().collect(Collectors.joining("\n"));
+                    }
+
+                } catch (MalformedURLException e) {
+                    e.printStackTrace();
+                    Log.e("WebView", "Fail 1", e);
+
+                } catch (IOException e) {
+                    e.printStackTrace();
+                    Log.e("WebView", "Fail 2", e);
+
+                }
+            }
+        }).start();
     }
 }
