+++ app/src/main/res/layout/fragment_notifications.xml
@@ -30,10 +30,16 @@
     <SeekBar
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:id="@+id/sbTextSize"
-        android:max="300"
-        android:progress="100"
+        android:id="@+id/seekBar"
+        android:max="100"
+        android:progress="50"
         />
+
+    <WebView
+        android:id="@+id/webView"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content" />
+
